api/requirements.txt
Flask==2.3.2
boto3==1.28.0
paho-mqtt==1.6.1
pyserial==3.5